Fix repository and addition assignments in _versioncontrol_amend_operations() by only assigning to operations of the currently processed VCS backend. This fix makes multiple backends work together also in the commit log.
As a welcome side effect - now that I'm sure that indexes don't affect
the sorting of arrays - we get a vc_op_id as array key for commit,
branch op and tag op arrays. Which is source compatible, as users that
don't know about the key (like now) must not have assumed what it contains,
they only may do so from now on.