API change! versioncontrol_get_current_item_tag() now returns just the tag name as string instead of the whole tag operation. This change needs to happen because otherwise items coming from get_all_item_tags() (as opposed to get_tagged_items() which already supplies the tag op) would have a very hard time.
That shouldn't hurt existing modules (outside the versioncontrol project,
at least) as the function was not yet used and was only implemented by
the CVS backend, I'll update that one at once.

Also, mark mandatary functions as such in the FakeVCS example backend.
Makes it easier to find those functions that have to be implemented.