D6 upgrade task 1.: Entirely new menu system.
...and yup, that did some heavy changes. Most notably, we're finally able
to get rid of that hook_user() category hack, which is great. I'm pretty
sure that it doesn't work the way it stands now (I can tell *that* much
by looking at the user.module sources), and most probably we'll need some
hook_menu_alter() magic, but the bulk of the menu conversion is done.

This was probably the most unwieldy chunk of the D6 port, everything else
should be more or less straightforward :)