Remove commit_restrictions submodule.
It is out of date now. That module was assuming that the hook scripts
will be calling the old hook_versioncontrol_write_access(), that is not
used anymore.

In the other side vcs_auth repository plugins are now kind of overlapping
with it, so let's remove this, and probably re-introduce in the future if
we want to support it again.