Change VersioncontrolRepositoryLogParserInterface to VersioncontrolRepositoryHistorySynchronizerInterface.
This is an attempt at replacing the old naming convention of "log
parsing" with "history synchronization". The former name derives from
the fact that the action tends to be "parse the output of <vcs> log",
which worked well enough as an internal name, but is really quite
misleading for new people coming in. "History synchronization" is better
as it speaks to the overall action, and also provides a better
conceptual framework for the methods attached to the interface itself
(which are now pretty much all "sync*" methods).